requirements-dev.txt
ruff
paho-mqtt==2.1.0
PyYAML==6.0.3

# Testing dependencies
pytest>=8.0.0
pytest-cov>=4.1.0
pytest-mock>=3.12.0
pytest-asyncio>=0.23.0
